Custodian v0.0.1 Custodian.Bots

The Bots context provides a boundary into the Custodian.Bots.Bot schema. It provides functions for listing, creating, updating, and deleting bots.

Functions

change_bot(bot)

change_bot(Custodian.Bots.Bot.t()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking bot changes.

Examples

iex> change_bot(bot)
%Ecto.Changeset{source: %Bot{}}

create_bot(attrs \\ %{})

create_bot(map()) :: Custodian.Bots.Bot.t()

Creates a bot.

Examples

iex> create_bot(%{field: value})
{:ok, %Bot{}}

iex> create_bot(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_bot(bot)

delete_bot(Custodian.Bots.Bot.t()) :: Custodian.Bots.Bot.t()

Deletes a Bot.

Examples

iex> delete_bot(bot)
{:ok, %Bot{}}

iex> delete_bot(bot)
{:error, %Ecto.Changeset{}}

get_bot!(id)

get_bot!(String.t()) :: Custodian.Bots.Bot.t()

Gets a single bot.

Raises Ecto.NoResultsError if the bot does not exist.

Examples

iex> get_bot!(123)
%Bot{}

iex> get_bot!(456)
** (Ecto.NoResultsError)

get_bot_by!(clauses)

get_bot_by!(Keyword.t()) :: Custodian.Bots.Bot.t()

Gets a single bot by some clause.

Raises Ecto.NoResultsError if the bot does not exist.

Examples

iex> get_bot_by!(repo_id: 123)
%Bot{}

iex> get_bot_by!(repo_id: 456)
** (Ecto.NoResultsError)

list_bots()

list_bots() :: [Custodian.Bots.Bot.t()]

Returns the list of bots.

Examples

iex> list_bots()
[%Bot{}, ...]

update_bot(bot, attrs)

update_bot(Custodian.Bots.Bot.t(), map()) :: Custodian.Bots.Bot.t()

Updates a bot.

Examples

iex> update_bot(bot, %{field: new_value})
{:ok, %Bot{}}

iex> update_bot(bot, %{field: bad_value})
{:error, %Ecto.Changeset{}}