Custodian v0.0.1 Custodian.Github.Client behaviour

Behaviour interface for taking actions on specific repos on GitHub.

Implemented using Tentacat in Custodian.Github.Tentacat.Client. In tests, implemented as a mock in Mockcat.

Callbacks

app()

app() :: struct()

Uses the configured private key to generate a JWT.

In development, the key is loaded from config/github_key.pem. In test, it’s loaded from test/support/github_key.pem. In production this is loaded from the GITHUB_API_KEY ENV var.

This token is used to generate access tokens that can be used for repo actions. The JWT can only be used for generating new tokens.

installation(installation_id)

installation(installation_id :: integer()) :: struct()

Uses the app JWT to generate an access token.

This token can be used to take action on specific repos.