Custodian v0.0.1 Custodian.Github.Labels behaviour

Behaviour interface for interacting with labels on pull requests on GitHub.

Implemented using Tentacat in Custodian.Github.Tentacat.Labels. In tests, implemented as a mock in Mockcat.

Callbacks

add(pull_request, label)

add(pull_request :: Custodian.Github.pull_request(), label :: String.t()) :: Custodian.Github.pull_request()

add(pull_request :: Custodian.Github.pull_request(), labels :: [String.t()]) :: Custodian.Github.pull_request()

Adds the provided label(s) to the given pull request.

all(pull_request)

all(pull_request :: Custodian.Github.pull_request()) :: [String.t()]

Returns a list of strings representing all of the labels on a given pull request.

remove(pull_request, label)

remove(pull_request :: Custodian.Github.pull_request(), label :: String.t()) :: Custodian.Github.pull_request()

Removes a given label from a given pull request.