Custodian v0.0.1 Custodian.Github.Tentacat.Merges

Provides an implementation of the Custodian.Github.Merges behaviour with the GitHub API.

This module provides a function to merge branches via the GitHub v3 API over HTTPS.

Functions

merge(bot, head, base)

Merges commits from base into head on a repo.

More info.

Examples

iex> merge(%Bot{}, "pr-1", "master")
{201, ""}