Custodian v0.0.1 Custodian.Github.Tentacat.References

Provides an implementation of the Custodian.Github.References behaviour with the GitHub API.

This module provides a function to delete branches via the GitHub v3 API over HTTPS.

Functions

remove(bot, branch)

remove(Custodian.Bots.Bot.t(), String.t()) :: tuple()

Removes the branch from the repo connected to bot.

More info.

Examples

iex> remove(%Bot{}, "pr-1")
{201, ""}