Custodian v0.0.1 CustodianWeb.Router.Helpers

Module with named helpers generated from CustodianWeb.Router.

Functions

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

webhook_path(conn_or_endpoint, action)

webhook_path(conn_or_endpoint, action, params)

webhook_url(conn_or_endpoint, action)

webhook_url(conn_or_endpoint, action, params)